Main activity backend other than seekbars
Made the images buttons so that they can go to the select avatar intent. Also go the images back into the main activity so that they display. And worked on sending the data from the main page to the display page. Have not inserted any data for the display page yet.